+++ src/app/student/course/course.component.ts
@@ -191,14 +191,29 @@ delete(id: string) {
       });
       return;
     }
-    this._courseService.deleteCourse(id).subscribe(() => {
-      this.getCoursesList();
-      Swal.fire({
-        title: 'Success',
-        text: 'Course deleted successfully.',
-        icon: 'success',
-      });
+
+    Swal.fire({
+      title: "Confirm Deletion",
+      text: "Are you sure you want to delete?",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#d33",
+      cancelButtonColor: "#3085d6",
+      confirmButtonText: "Delete",
+      cancelButtonText: "Cancel",
+    }).then((result) => {
+      if (result.isConfirmed){
+        this._courseService.deleteCourse(id).subscribe(() => {
+          this.getCoursesList();
+          Swal.fire({
+            title: 'Success',
+            text: 'Course deleted successfully.',
+            icon: 'success',
+          });
+        });
+      }
     });
+   
   });
 }
